chore: resolve comment 1
