+++ packages/ui/src/components/data-formater/dataformater.tsx
@@ -1,0 +1,9 @@
+// Função auxiliar para formatar a data no formato "yyyy-MM-ddTHH:mm"
+export const formatDateTimeLocal = (date: Date) => {
+	const year = date.getFullYear();
+	const month = String(date.getMonth() + 1).padStart(2, "0");
+	const day = String(date.getDate()).padStart(2, "0");
+	const hours = String(date.getHours()).padStart(2, "0");
+	const minutes = String(date.getMinutes()).padStart(2, "0");
+	return `${year}-${month}-${day}T${hours}:${minutes}`;
+};
